Finish backend refactor
