+++ server-super-turtle-tracker/README.md
@@ -1,4 +1,4 @@
 # TurtleTrackerBackend
 
-To test, run `node index.js`. Connect to the server using `localhost:<port>`.
+To test, run `node server.js`. Connect to the server using `localhost:<port>`.
 To deploy, run `git subtree push --prefix server-super-turtle-tracker heroku master`.
