+++ server-super-turtle-tracker/package.json
@@ -29,6 +29,7 @@
     "firebase-admin": "^9.1.1",
     "fs": "0.0.1-security",
     "fs-extra": "^8.1.0",
+    "http-status": "^1.4.2",
     "mongodb": "^3.4.0",
     "multer": "^1.4.2",
     "pg": "^7.12.1",
