+++ server-super-turtle-tracker/src/db/queries/photoQueries.ts
