--- server-super-turtle-tracker/src/helpers/status.ts
@@ -1,13 +1,0 @@
-export const successMessage = { status: "success" };
-
-export const errorMessage = { status: "error" };
-
-export const status = {
-  success: 200,
-  error: 500,
-  notfound: 404,
-  unauthorized: 401,
-  created: 201,
-  bad: 400,
-  conflict: 409
-};
