+++ server-super-turtle-tracker/src/middlewares/auth.ts
@@ -1,0 +1,28 @@
+import * as dotenv from "dotenv";
+import express from "express";
+import status = require("http-status");
+
+dotenv.config();
+
+const SECRET = process.env.BACKEND_SECRET;
+
+export default (
+  request: express.Request,
+  response: express.Response,
+  next: express.NextFunction,
+): void => {
+  let token = " ";
+  if (request.headers.authorization !== undefined) {
+    try {
+      const [, tok] = request.headers.authorization.split(" ");
+      token = tok;
+    } catch (error) {
+      response.status(status.UNAUTHORIZED).json("No Bearer Token given");
+    }
+  }
+  if (token === SECRET) {
+    next();
+  } else {
+    response.status(status.UNAUTHORIZED).json("Unauthorized Request");
+  }
+};
