+++ server-super-turtle-tracker/src/routes/mailRoutes.ts
@@ -1,0 +1,9 @@
+import express from "express";
+import verifyAuth from "../middlewares/auth";
+import sendMail from "../services/mailService";
+
+const router = express.Router();
+
+router.get("/:address", verifyAuth, sendMail);
+
+export default router;
